pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hierarchical-contrastive-qa-with-adversarial-unanswerable-detection"
version = "0.1.0"
description = "Hierarchical contrastive QA with adversarial training for unanswerable detection on SQuAD 2.0"
authors = [{name = "Alireza Shojaei"}]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "sentence-transformers>=2.2.0",
    "datasets>=2.12.0",
    "evaluate>=0.4.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.2.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "mlflow>=2.3.0",
    "tensorboard>=2.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
]

[project.scripts]
hcqa-train = "hierarchical_contrastive_qa_with_adversarial_unanswerable_detection.cli.train:main"
hcqa-evaluate = "hierarchical_contrastive_qa_with_adversarial_unanswerable_detection.cli.evaluate:main"
hcqa-predict = "hierarchical_contrastive_qa_with_adversarial_unanswerable_detection.cli.predict:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
